+++ Editor/Tiles/AutoTile/UI/AutoTileEditorElement.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Unity.Collections;
 using UnityEditor.UIElements;
 using UnityEngine;
 using UnityEngine.Tilemaps;
@@ -15,8 +16,20 @@ namespace UnityEditor.Tilemaps
 
         private static readonly float s_MaxSliderScale = 2.5f;
 
+        private static class Styles
+        {
+            public static readonly string defaultSpriteTooltip = L10n.Tr("The Sprite set when there are no matches.");
+            public static readonly string defaultGameObjectTooltip = L10n.Tr("The GameObject instantiated when set on the Tilemap.");
+            public static readonly string tileColliderTooltip = L10n.Tr("The Collider Type used for generating colliders.");
+            public static readonly string maskTypeTooltip = L10n.Tr("Mask Type for setting Rules for the AutoTile. Use 2x2 for a 16 Sprite ruleset and 3x3 for a 47 Sprite ruleset.");
+            public static readonly string randomTooltip = L10n.Tr("Randomly picks a Sprite if multiple Sprites share the same mask. Otherwise, uses the first Sprite set with the mask.");
+            public static readonly string physicsShapeCheckTooltip = L10n.Tr("Checks whether the Sprite used has a physics shape. If not, the Collider Type will be set to None.");
+        }
+
         private ListView m_TextureList;
         private ScrollView m_TextureScroller;
+
+        private Toggle m_PhysicsShapeCheckToggle;
 
         private Dictionary<Texture2D, AutoTileTextureSource> textureToElementMap =
             new Dictionary<Texture2D, AutoTileTextureSource>();
@@ -39,22 +52,39 @@ namespace UnityEditor.Tilemaps
             var defaultSprite = new ObjectField("Default Sprite");
             defaultSprite.objectType = typeof(Sprite);
             defaultSprite.bindingPath = "m_DefaultSprite";
+            defaultSprite.tooltip = Styles.defaultSpriteTooltip;
             defaultProperties.Add(defaultSprite);
 
             var defaultGameObject = new ObjectField("Default GameObject");
             defaultGameObject.objectType = typeof(GameObject);
             defaultGameObject.bindingPath = "m_DefaultGameObject";
+            defaultGameObject.tooltip = Styles.defaultGameObjectTooltip;
             defaultProperties.Add(defaultGameObject);
 
             var tileColliderType = new EnumField("Tile Collider");
             tileColliderType.bindingPath = "m_DefaultColliderType";
+            tileColliderType.tooltip = Styles.tileColliderTooltip;
+            tileColliderType.RegisterValueChangedCallback(ColliderTypeChanged);
             defaultProperties.Add(tileColliderType);
+
+            m_PhysicsShapeCheckToggle = new Toggle("Has Physics Shape");
+            m_PhysicsShapeCheckToggle.name = m_PhysicsShapeCheckToggle.label;
+            m_PhysicsShapeCheckToggle.bindingPath = "m_PhysicsShapeCheck";
+            m_PhysicsShapeCheckToggle.tooltip = Styles.physicsShapeCheckTooltip;
+            defaultProperties.Add(m_PhysicsShapeCheckToggle);
 
             var maskType = new EnumField("Mask Type");
             maskType.bindingPath = "m_MaskType";
+            maskType.tooltip = Styles.maskTypeTooltip;
             maskType.RegisterValueChangedCallback(MaskTypeChanged);
-
             defaultProperties.Add(maskType);
+
+            var random = new Toggle("Random");
+            random.name = random.label;
+            random.bindingPath = "m_Random";
+            random.tooltip = Styles.randomTooltip;
+            random.RegisterValueChangedCallback(RandomChanged);
+            defaultProperties.Add(random);
 
             Add(defaultProperties);
 
@@ -84,6 +114,8 @@ namespace UnityEditor.Tilemaps
             if (autoTile == null)
                 return;
 
+            m_PhysicsShapeCheckToggle.SetEnabled(autoTile.m_DefaultColliderType == Tile.ColliderType.Sprite);
+
             m_TextureList.itemsSource = m_AutoTile.m_TextureList;
             m_TextureList.Rebuild();
             m_TextureList.RefreshItems();
@@ -106,7 +138,7 @@ namespace UnityEditor.Tilemaps
                     if (textureToElementMap.TryGetValue(spriteTexture, out var at))
                     {
                         at.InitialiseSpriteMask(sprite, mask);
-                        if (isDuplicate && mask > 0)
+                        if (!autoTile.random && isDuplicate && mask > 0)
                             at.SetDuplicate(sprite, true);
                     }
                 }
@@ -218,11 +250,12 @@ namespace UnityEditor.Tilemaps
                 he.Add(saveButton);
 
                 var minLength = Math.Max(texture2D.width, texture2D.height);
-                var start = 256.0f / minLength;
-
+                var start = 0.5f;
+                if (minLength > 512.0f)
+                    start = 256.0f / minLength;
                 var sliderValue = Math.Min(Mathf.Max(start, m_AutoTile.m_TextureScaleList[i]), s_MaxSliderScale);
-
                 var slider = new Slider("Scale", start, s_MaxSliderScale, SliderDirection.Horizontal, 0.1f);
+                slider.name = "ScaleSlider";
                 slider.style.flexGrow = 0.9f;
                 slider.value = Mathf.Max(start, sliderValue);
                 slider.userData = i;
@@ -250,7 +283,7 @@ namespace UnityEditor.Tilemaps
             if (oldMask != 0)
             {
                 var spriteList = autoTile.m_AutoTileDictionary[oldMask].spriteList;
-                if (spriteList.Count > 2)
+                if (!autoTile.random && spriteList.Count > 2)
                 {
                     if (textureToElementMap.TryGetValue(sourceTexture, out var at))
                     {
@@ -258,7 +291,7 @@ namespace UnityEditor.Tilemaps
                     }
                 }
 
-                if (spriteList.Count == 2)
+                if (!autoTile.random &&spriteList.Count == 2)
                 {
                     foreach (var autoTileSprite in spriteList)
                     {
@@ -273,7 +306,7 @@ namespace UnityEditor.Tilemaps
             autoTile.RemoveSprite(sprite, oldMask);
             autoTile.AddSprite(sprite, sourceTexture, newMask);
 
-            if (newMask != 0)
+            if (newMask != 0 && !autoTile.random)
             {
                 var spriteList = autoTile.m_AutoTileDictionary[newMask].spriteList;
                 if (spriteList.Count < 2)
@@ -289,24 +322,48 @@ namespace UnityEditor.Tilemaps
             }
         }
 
+        private void ColliderTypeChanged(ChangeEvent<Enum> evt)
+        {
+            m_PhysicsShapeCheckToggle.SetEnabled((Tile.ColliderType) evt.newValue == Tile.ColliderType.Sprite);
+        }
+
+        private void RandomChanged(ChangeEvent<bool> evt)
+        {
+            TexturesChanged();
+        }
+
+        private void UpdateTextureList()
+        {
+            m_TextureList.Rebuild();
+            TexturesChanged();
+        }
+
         private void ItemListAdded(IEnumerable<int> insertions)
         {
             // Note: m_AutoTile.m_TextureList is increased before this method
             foreach (var i in insertions)
                 m_AutoTile.m_TextureScaleList.Insert(i, AutoTile.s_DefaultTextureScale);
             SaveTile();
-            m_TextureList.Rebuild();
-            TexturesChanged();
+            m_TextureList.schedule.Execute(UpdateTextureList);
         }
 
         private void ItemListRemoved(IEnumerable<int> removals)
         {
             // Note: m_AutoTile.m_TextureList is reduced after this method ends
+            int count = 0;
+            NativeArray<int> removalNative = new NativeArray<int>(m_AutoTile.m_TextureScaleList.Count, Allocator.Temp);
             foreach (var i in removals)
-                m_AutoTile.m_TextureScaleList.RemoveAt(i);
+            {
+                removalNative[count++] = i;
+            }
+            for (var idx = count - 1; idx >= 0; idx--)
+            {
+                m_AutoTile.m_TextureScaleList.RemoveAt(removalNative[idx]);
+            }
+            removalNative.Dispose();
+
             SaveTile();
-            m_TextureList.Rebuild();
-            TexturesChanged();
+            m_TextureList.schedule.Execute(UpdateTextureList);
         }
 
         private void TexturesChanged()