+++ Runtime/AssemblyInfo.cs
@@ -3,3 +3,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.2D.Tilemap.Extras.Editor")]
 [assembly: InternalsVisibleTo("Unity.2D.Tilemap.Extras.Tests")]
 [assembly: InternalsVisibleTo("Unity.2D.Tilemap.Extras.EditorTests")]
+[assembly: InternalsVisibleTo("Unity.2D.Tilemap.Extras.RuntimeTests")]